Changed 'context.jsx' to 'AuthContext.jsx' and moved into parent folder named 'context'
