+++ src/components/layouts/ApplicationLayout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom"
-import { AuthProvider } from "../../Context"
+import { AuthProvider } from "../../context/AuthContext"
 
 const ApplicationLayout = () => {
     return <>
